chore: extract oauth module to a single directory
